# Optimize dired-all-the-icons by unlimiting memoization limit