@ymarco 2020-10-27
I think I'll be making a PR to all-the-icons to change the cache after all.