@ymarco 2020-10-27
https://github.com/domtronn/all-the-icons.el/pull/227